requirements.txt
customtkinter
pyautogui
pytesseract
Pillow
keyboard
requests